Rename HexInput to HexColorInput
