+++ src/index.ts
@@ -1,11 +1,13 @@
-// Public components
+// Color picker components
 export { HexColorPicker } from "./components/HexColorPicker";
-export { HexColorInput } from "./components/common/HexInput";
 export { HslColorPicker } from "./components/HslColorPicker";
 export { HslStringColorPicker } from "./components/HslStringColorPicker";
 export { HsvColorPicker } from "./components/HsvColorPicker";
 export { RgbColorPicker } from "./components/RgbColorPicker";
 export { RgbStringColorPicker } from "./components/RgbStringColorPicker";
+
+// Additional components
+export { HexColorInput } from "./components/HexColorInput";
 
 // Color model types
 export { RGB, HSL, HSV } from "./types";
